Further updates re TeamLandingPage
- Added Style option to Typography file for further customization
- Changed TeamLandingPageTables to TeamLandingPageElements.
- Began moving the components I have built on TeamLandingPage into
TeamLandingPageElements to keep the master file (TeamLandingPage) clear
from too much code.
- Added top players functionality which will be used in TeamLandingPage,
however I still need to work on logic to prevent reveal of CFB players
overall numeric value when switching from CFB to NFL.
